Refactor monster interaction: Moved Monster Fighting into Base Monster Room Class
